requirements.txt
Django
django-allauth
django-cors-headers
django-phonenumber-field
djangorestframework
dj-database-url
phonenumbers
pillow
pycparser
python-dotenv
pytz
sqlparse
tzdata
psycopg2
gunicorn
whitenoise
djangorestframework
python-telegram-bot
requests
cryptography
pyyaml
ua-parser
user-agents
pandas==2.3.2
django-filter==25.1
openpyxl==3.1.5
cloudinary==1.44.1
requests-toolbelt

